Cleanup import statements
