+++ src/pycel/lib/function_info.py
@@ -1,4 +1,5 @@
 import os
+
 from pycel.lib.function_info_data import function_info
 
 # set of all function names
