+++ tox.ini
@@ -15,18 +15,23 @@ deps =
     pytest
     pytest-flake8
     pytest-cov
+    codecov
+    flake8-import-order
     restructuredtext-lint
+    tox-travis
 
 commands =
     pytest --flake8 --cov {envsitepackagesdir}/pycel --cov-report term-missing
 
 [flake8]
 
+application-import-names = pycel
+
 max-line-length = 80
 
 import-order-style = smarkets
 
-select = C,D300,E,F,I1,W
+select = C,D300,E,F,I,W
 
 [pytest]
